+++ image/build_os2borgerpc_image.sh
@@ -10,13 +10,21 @@ printf "\n\n%s\n\n" "===== RUNNING: $0 ====="
 
 # Handle optional arguments
 COUNT=0
-if [ "$1" = "--clean" ] || [ "$2" = "--clean" ]; then
+if [ "$1" = "--clean" ] || [ "$2" = "--clean" ] || [ "$3" = "--clean" ]; then
   CLEAN_BUILD=true
+  echo "Arguments: Clean building"
   COUNT=$((COUNT+1))
 fi
 
-if [ "$1" == "--skip-build-deps" ] || [ "$2" == "--skip-build-deps" ]; then
+if [ "$1" == "--skip-build-deps" ] || [ "$2" == "--skip-build-deps" ] || [ "$3" == "--skip-build-deps" ]; then
   SKIP_BUILD_DEPS=true
+  echo "Arguments: Skipping installing build deps"
+  COUNT=$((COUNT+1))
+fi
+
+if [ "$1" == "--lang-all" ] || [ "$2" == "--lang-all" ] || [ "$3" == "--lang-all" ]; then
+  LANG_ALL=true
+  echo "Arguments: Setting up better support for all languages"
   COUNT=$((COUNT+1))
 fi
 # Remove the optional arguments if there were any, so we only have the required ones left
@@ -28,10 +36,11 @@ IMAGE_NAME=$2
 
 if [[ -z $ISO_PATH || -z $IMAGE_NAME ]]
 then
-    echo "Usage: "$0"  [--clean] [--skip-build-deps] iso_file image_name"
+    echo "Usage: "$0"  [--clean] [--skip-build-deps] [--lang-all] iso_file image_name"
     echo ""
-    echo "--clean: pass this argument to first delete temp build files, e.g. from within iso/"
-    echo "--skip-build-deps: pass this argument to skip installing build dependencies. Useful if testing on a non-debian-system, ie. without apt"
+    echo "--clean: First delete temp build files, e.g. from within iso/"
+    echo "--skip-build-deps: Skip installing build dependencies. Useful if testing on a non-debian-system, ie. without apt"
+    echo "--lang-all: Build an image with more multi-language-support out of the box"
     echo "iso_file must be a valid path to the ISO file to be remastered"
     echo "image_name is the name of the output image"
     echo ""
@@ -47,7 +56,7 @@ unmount_cleanup() {
 
 figlet "Building OS2borgerPC"
 
-echo "Ignore errors about zsys daemon in the log output"
+echo "You can ignore errors about zsys daemon in the log output"
 
 set -ex
 
@@ -56,7 +65,7 @@ then
     # In case it was cancelled prematurely and /tmp is still bind-mounted to squashfs-root/tmp
     unmount_cleanup
     sudo chattr -f -i squashfs-root/var/lib/lightdm/.cache/unity-greeter/state || true
-    sudo rm -rf iso/.disk/ iso/* squashfs squashfs-root/ /tmp/build_installed_packages_list.txt /tmp/scripts_installed_packages_list.txt /tmp/os2borgerpc_install_log.txt /tmp/os2borgerpc_upgrade_log.txt boot_hybrid.img ubuntu22-desktop-amd64.efi
+    sudo rm --recursive --force iso/.disk/ iso/* squashfs squashfs-root/ /tmp/build_installed_packages_list.txt /tmp/scripts_installed_packages_list.txt /tmp/os2borgerpc_install_log.txt /tmp/os2borgerpc_upgrade_log.txt boot_hybrid.img ubuntu22-desktop-amd64.efi
 fi
 
 if [ ! "$SKIP_BUILD_DEPS" ]
@@ -67,18 +76,18 @@ fi
 build/extract_iso.sh "$ISO_PATH" iso
 
 # Unsquash and customize
-sudo unsquashfs -f iso/casper/filesystem.squashfs > /dev/null
+sudo unsquashfs -force iso/casper/filesystem.squashfs > /dev/null
 
 # Mounting in our own tmp into the chroot so we retain access to the log files written from within it
 # This will fail in the pipeline due to a permissions error, but it will work locally
 sudo mount --bind /tmp $TMP_MOUNT_POINT || true
 
 echo "Copying in image building files before chrooting"
-sudo rsync -r --exclude squashfs-root --exclude image/*.iso --exclude .git --exclude image/iso ../ $BUILD_FILES_COPY_DESTINATION
+sudo rsync --recursive --exclude squashfs-root --exclude image/*.iso --exclude .git --exclude image/iso ../ $BUILD_FILES_COPY_DESTINATION
 
 figlet "About to enter chroot"
 
-build/chroot_os2borgerpc.sh squashfs-root ./build/prepare_os2borgerpc.sh
+build/chroot_os2borgerpc.sh squashfs-root ./build/prepare_os2borgerpc.sh "$LANG_ALL"
 
 
 # Regenerate manifest
@@ -86,23 +95,33 @@ build/chroot_os2borgerpc.sh squashfs-root build/create_manifest.sh > iso/casper/
 figlet "Exiting chroot"
 
 cp iso/casper/filesystem.manifest iso/casper/filesystem.manifest-desktop
-sed -i '/ubiquity/d' iso/casper/filesystem.manifest-desktop
-sed -i '/casper/d' iso/casper/filesystem.manifest-desktop
+sed --in-place '/ubiquity/d' iso/casper/filesystem.manifest-desktop
+sed --in-place '/casper/d' iso/casper/filesystem.manifest-desktop
 
 
 # Build squashfs for the ISO
 
 # First delete the image building files from squashfs again
-sudo rm -rf $BUILD_FILES_COPY_DESTINATION/* squashfs-root/*.sh
+sudo rm --recursive --force $BUILD_FILES_COPY_DESTINATION/* squashfs-root/*.sh
 
 rm iso/casper/filesystem.squashfs
 sudo mksquashfs squashfs-root iso/casper/filesystem.squashfs
 
 # Calculate FS size
-printf $(sudo du -sx --block-size=1 squashfs-root | cut -f1) > iso/casper/filesystem.size
+printf $(sudo du --summarize --one-file-system --block-size=1 squashfs-root | cut --fields 1) > iso/casper/filesystem.size
 
 # Overwrite preseed etc.
-cp -r iso_overwrites/* iso/
+cp --recursive iso_overwrites/* iso/
+# Make a few changes to the copied preseed file if we're building an image with multi-language-support
+if [ "$LANG_ALL" ]
+then
+  # Don't preseed locale/timezone
+  sed --in-place --expression "/keyboard-configuration/d" --expression "\@debian-installer/locale@d" --expression "\@d-i time/zone@d" iso/preseed/ubuntu.seed
+  # Don't use the OS2-specific background image
+  rm iso/boot/grub/borgerpc_grub_bg.png
+  # TODO: Potentially remove in the future if converting to a true language agnostic image: Hardcode ISO bootup text to Swedish
+  sed --in-place --expression "s/Install OS2borgerPC from this medium/Installera Sambruk MedborgarPC från detta medium/" iso/boot/grub/grub.cfg
+fi
 
 # Recalculate MD5 sums.
 cd iso
@@ -122,12 +141,12 @@ dd if="$ISO_PATH" bs=1 count=446 of=$mbr
 
 # Extract EFI partition image
 
-skip=$(/sbin/fdisk -l "$ISO_PATH" | grep -F '.iso2 ' | awk '{print $2}')
+skip=$(/sbin/fdisk --list "$ISO_PATH" | grep --fixed-strings '.iso2 ' | awk '{print $2}')
 
-size=$(/sbin/fdisk -l "$ISO_PATH" | grep -F '.iso2 ' | awk '{print $4}')
+size=$(/sbin/fdisk --list "$ISO_PATH" | grep --fixed-strings '.iso2 ' | awk '{print $4}')
 
 dd if="$ISO_PATH" bs=512 skip="$skip" count="$size" of=$efi
 
 # Make image
 
-xorriso -as mkisofs -r   -V "$IMAGE_NAME" -o "$IMAGE_NAME".iso   -J -joliet-long -l -iso-level 3 -partition_offset 16 --grub2-mbr $mbr --mbr-force-bootable -append_partition 2 0xEF $efi -appended_part_as_gpt -c boot.catalog -b boot/grub/i386-pc/eltorito.img -no-emul-boot   -boot-load-size 4 -boot-info-table --grub2-boot-info  -eltorito-alt-boot -e '--interval:appended_partition_2:all::' -no-emul-boot iso
+xorriso -as mkisofs -r -V "$IMAGE_NAME" -o "$IMAGE_NAME".iso -J -joliet-long -l -iso-level 3 -partition_offset 16 --grub2-mbr $mbr --mbr-force-bootable -append_partition 2 0xEF $efi -appended_part_as_gpt -c boot.catalog -b boot/grub/i386-pc/eltorito.img -no-emul-boot   -boot-load-size 4 -boot-info-table --grub2-boot-info  -eltorito-alt-boot -e '--interval:appended_partition_2:all::' -no-emul-boot iso